+++ landing_page_checker/forms.py
@@ -1,5 +1,4 @@
-from directory.models import SecuredropOwner
-from landing_page_checker.models import SecuredropPage
+from directory.models import SecuredropOwner, SecuredropPage
 from django.contrib.auth import get_user_model
 from django import forms
 from django.core.exceptions import ValidationError
