+++ landing_page_checker/management/commands/scan.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand, CommandError
 
-from landing_page_checker.models import SecuredropPage
+from directory.models import SecuredropPage
 from landing_page_checker.landing_page.scanner import bulk_scan
 from landing_page_checker.utils import url_to_domain
