+++ landing_page_checker/views.py
@@ -8,9 +8,8 @@ from django_otp.decorators import otp_required
 
 from common.decorators import directory_management_required
 from common.models.settings import DirectorySettings
-from directory.models import DirectoryPage
+from directory.models import DirectoryPage, SecuredropPage
 from landing_page_checker.forms import SecuredropPageOwnerForm
-from landing_page_checker.models import SecuredropPage
 
 
 class SecuredropListView(ListView):
